requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
aiocache==0.12.2
sqlalchemy==2.0.30
python-dotenv==1.0.1
apscheduler==3.10.4
gspread==6.1.2
oauth2client==4.1.3